feat(refactor): change package for emails template from .service.template to service.email.template
