+++ memory-bank/progress.md
@@ -1,5 +1,40 @@
 # Implementation Progress
 
+## System wysyłki emaili - Zakończony [2025-02-28]
+
+### Work Done
+- Przeprowadzono analizę obecnych szablonów emaili
+- Zaprojektowano nowe, rozbudowane szablony z HTML i CSS
+- Udokumentowano decyzję o rozszerzeniu szablonów w decisionLog.md
+- Zaimplementowano system szablonów:
+  - Utworzono bazową klasę EmailTemplate
+  - Zaimplementowano system zmiennych TemplateVariables
+  - Utworzono szablony dla wszystkich typów emaili
+  - Zmodyfikowano EmailService aby korzystał z nowych szablonów
+  - Zaktualizowano testy
+  - Dodano konfigurację app.base-url
+
+### Next Steps
+1. Rozważyć dodanie nowych typów emaili:
+   - Powiadomienia o logowaniu
+   - Powiadomienia o zmianie ustawień konta
+   - Newsletter z aktualnościami
+
+2. Potencjalne ulepszenia:
+   - Dodanie opcji wyboru języka w szablonach
+   - Implementacja systemu śledzenia otwarć emaili
+   - Dodanie opcji personalizacji kolorystyki przez użytkownika
+
+3. Testy:
+   - Przetestować wyświetlanie emaili w różnych klientach pocztowych
+   - Zweryfikować zachowanie responsywnego designu
+   - Sprawdzić obsługę różnych zestawów znaków
+
+4. Dokumentacja:
+   - Przygotować dokumentację dla developerów dot. tworzenia nowych szablonów
+   - Utworzyć przewodnik stylów dla szablonów emaili
+   - Udokumentować system zmiennych do personalizacji
+
 ## System Resetowania Hasła - Zakończony [2025-02-28]
 
 ### Completed Tasks
