+++ src/main/java/info/mackiewicz/bankapp/service/email/template/EmailContent.java
@@ -1,4 +1,4 @@
-package info.mackiewicz.bankapp.service.template;
+package info.mackiewicz.bankapp.service.email.template;
 
 /**
  * Data class for email content.
