+++ src/test/java/info/mackiewicz/bankapp/config/TestConfig.java
@@ -1,7 +1,8 @@
 package info.mackiewicz.bankapp.config;
 
 import info.mackiewicz.bankapp.service.email.EmailSender;
-import info.mackiewicz.bankapp.service.template.EmailTemplateProvider;
+import info.mackiewicz.bankapp.service.email.template.EmailTemplateProvider;
+
 import org.mockito.Mockito;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
